Создано хранилище пиниа и декомпозирован код
