+++ src/App.vue
@@ -1,3 +1,4 @@
 <template>
+  <!-- динамическое окно для показа страницы -->
   <router-view />
 </template>
