--- src/stores/mainStore.js
@@ -1,8 +1,0 @@
-import { defineStore } from 'pinia'
-
-export const useMainStore = defineStore('main', {
-  state: () => ({
-    // здесь будет состояние
-  })
-})
-
